// frontend-legacy/src/client/types.gen.ts
export type StateValueType = 'boolean' | 'number' | 'string' | 'enum';
// ...
export type StoryStateVariableBase = {
    key: string;
    value_type?: StateValueType;
    default_value?: (unknown | null);
    enum_values?: (Array<(string)> | null);
    description?: (string | null);
    category?: (string | null);
};

export type StoryStateVariablePublic = {
    key: string;
    value_type?: StateValueType;
    default_value?: (unknown | null);
    enum_values?: (Array<(string)> | null);
    description?: (string | null);
    category?: (string | null);
    id: string;
    story_id: string;
    story_version: number;
};